server-b/requirements.txt
fastapi
uvicorn
slowapi
openai
pydantic
yfinance
akshare
httpx
python-dotenv
pyyaml
reportlab
pytest
redis
